# server/api/controllers/examples/controller.ts
import ExamplesService from '../../services/examples.service';
import { Request, Response, NextFunction } from 'express';
import * as HttpStatus from 'http-status-codes';
import HousesService from '../../services/houses.service';

export class Controller {

  async all(req: Request, res: Response, next: NextFunction) {
    try {

      const doc = await HousesService.getHouses();
      return res.status(HttpStatus.OK).json(doc);

      // const docs = await ExamplesService.all();
      // return res.status(HttpStatus.OK).json(docs);
    }
    catch (err) {
      return next(err);
    }
  }

  async byId(req: Request, res: Response, next: NextFunction) {
    try {
      const doc = await ExamplesService.byId(req.params.id);
      return res.status(HttpStatus.OK).json(doc);
    }
    catch (err) {
      return next(err);
    }
  }

  async create(req: Request, res: Response, next: NextFunction) {
    try {
      const doc = await ExamplesService.create(req.body);
      return res.status(HttpStatus.CREATED).location(`/api/v1/examples/${doc._id}`).json(doc);
    }
    catch (err) {
      return next(err);
    }
  }

  async patch(req: Request, res: Response, next: NextFunction) {
    try {
      const doc = await ExamplesService.patch(req.params.id, req.body);
      return res.status(HttpStatus.OK).location(`/api/v1/examples/${doc._id}`).json(doc);
    }
    catch (err) {
      return next(err);
    }
  }

  async remove(req: Request, res: Response, next: NextFunction) {
    try {
      const doc = await ExamplesService.remove(req.params.id);
      return res.status(HttpStatus.NO_CONTENT).send();
    }
    catch (err) {
      return next(err);
    }
  }

  async getHouses(req: Request, res: Response, next: NextFunction) {
    try {

      const doc = await HousesService.getHouses();
      return res.status(HttpStatus.OK).json(doc);
    }
    catch (err) {
      return next(err);
    }
  }

}

export default new Controller();