# server/common/swagger/Api.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: HH-NG
  description: HH Real Estates App
basePath: /api/v1
tags:
  - name: Examples
    description: Simple example endpoints
  - name: Specification
    description: The swagger API specification

consumes:
  - application/json
produces:
  - application/json


definitions:
  ExampleBody:
    type: object
    title: example
    required:
      - name
    properties:
      name:
        type: string
        example: no_stress
  HouseBody:
    type: object
    title: house
    required:
      - name
    properties:
      name:
        type: string
        house: no_stress

paths:
  /examples:
    get:
      tags:
        - Examples
      description: Fetch all examples
      responses:
        200:
          description: Returns all examples
    post:
      tags:
        - Examples
      description: Create a new example
      parameters:
        - name: example
          in: body
          description: an example
          required: true
          schema: 
            $ref: "#/definitions/ExampleBody"
      responses:
        200:
          description: Returns newly created example

  /examples/{id}:
    get:
      tags:
        - Examples
      description: Get an example
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the example to retrieve
          type: string
      responses:
        200:
          description: Return the example with the specified id
        404:
          description: Example not found
    patch:
      tags:
        - Examples
      description: Update an example
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the example to update
          type: string
        - name: example
          in: body
          description: example attribtues to update
          required: true
          schema: 
            $ref: "#/definitions/ExampleBody"
      responses:
        200:
          description: Return updated examples
        404:
          description: Example not found
    delete:
      tags:
        - Examples
      description: Delete an example
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the example to delete
          type: string
      responses:
        204:
          description: Delete the example with the specified id
        404:
          description: Example not found

  /houses:
    get:
      tags:
        - Houses
      description: Fetch all houses
      responses:
        200:
          description: Returns all houses
    post:
      tags:
        - Houses
      description: Create a new houses
      parameters:
        - name: house
          in: body
          description: a house
          required: true
          schema: 
            $ref: "#/definitions/HouseBody"
      responses:
        200:
          description: Returns newly created house
  
  /houses/{id}:
    get:
      tags:
        - Houses
      description: Get a house
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the house to retrieve
          type: string
      responses:
        200:
          description: Return the house with the specified id
        404:
          description: House not found
    patch:
      tags:
        - Houses
      description: Update a house
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the house to update
          type: string
        - name: house
          in: body
          description: house attribtues to update
          required: true
          schema: 
            $ref: "#/definitions/HouseBody"
      responses:
        200:
          description: Return updated houses
        404:
          description: House not found
    delete:
      tags:
        - Houses
      description: Delete a house
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the house to delete
          type: string
      responses:
        204:
          description: Delete the house with the specified id
        404:
          description: House not found

  /spec:
    get:
      tags:
        - Specification
      responses:
        200:
          description: Return the API specification